requirements.txt
ansi2html==1.8.0
astroid==3.0.1
autopep8==2.0.4
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dash==2.14.2
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
dill==0.3.7
docstring-to-markdown==0.13
flake8==6.1.0
Flask==3.0.0
fonttools==4.45.1
idna==3.6
importlib-metadata==6.8.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mccabe==0.7.0
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
pandas==2.1.3
parso==0.8.3
Pillow==10.1.0
platformdirs==4.0.0
plotly==5.18.0
pluggy==1.3.0
pycodestyle==2.11.1
pydocstyle==6.3.0
pyflakes==3.1.0
pylint==3.0.2
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
PyQt6-sip==13.6.0
pyqt6-tools==6.4.2.3.3
pyqtgraph==0.13.3
PySide6==6.6.0
PySide6-Addons==6.6.0
PySide6-Essentials==6.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.9.0
pytoolconfig==1.2.6
pytz==2023.3.post1
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
requests==2.31.0
retrying==1.3.4
rope==1.11.0
scipy==1.11.4
shiboken6==6.6.0
six==1.16.0
snowballstemmer==2.2.0
speedtest-cli==2.1.3
tenacity==8.2.3
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
urllib3==2.1.0
Werkzeug==3.0.1
whatthepatch==1.0.5
yapf==0.40.2
zipp==3.17.0
